backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.20
pandas==2.2.3
numpy==2.2.4
chainladder==0.8.24
python-docx==1.1.2
xlsxwriter==3.2.2
openai==1.76.0
supabase==2.15.0
matplotlib==3.10.1
